refactor: deduplicate resolve_type_name and address review comments
