+++ crates/server/src/handlers/pallets/events.rs
@@ -7,7 +7,7 @@
 
 use crate::handlers::pallets::common::{
     AtResponse, PalletError, PalletItemQueryParams, PalletQueryParams, RcBlockFields,
-    RcPalletItemQueryParams, RcPalletQueryParams, resolve_block_for_pallet,
+    RcPalletItemQueryParams, RcPalletQueryParams, resolve_block_for_pallet, resolve_type_name,
 };
 use crate::state::AppState;
 use crate::utils;
@@ -351,54 +351,6 @@ fn find_pallet<'a>(
     metadata
         .pallets()
         .find(|pallet| pallet.name().to_lowercase() == pallet_id_lower)
-}
-
-fn resolve_type_name(types: &scale_info::PortableRegistry, type_id: u32) -> String {
-    if let Some(ty) = types.resolve(type_id) {
-        if let scale_info::TypeDef::Variant(v) = &ty.type_def {
-            let is_simple_enum = v.variants.iter().all(|var| var.fields.is_empty());
-            if is_simple_enum {
-                let variant_names: Vec<String> = v
-                    .variants
-                    .iter()
-                    .map(|var| format!("\"{}\"", var.name))
-                    .collect();
-                return format!("{{\"_enum\":[{}]}}", variant_names.join(","));
-            }
-        }
-
-        if !ty.path.segments.is_empty() {
-            return ty.path.segments.last().unwrap().clone();
-        }
-        match &ty.type_def {
-            scale_info::TypeDef::Primitive(p) => format!("{:?}", p).to_lowercase(),
-            scale_info::TypeDef::Compact(c) => {
-                format!("Compact<{}>", resolve_type_name(types, c.type_param.id))
-            }
-            scale_info::TypeDef::Sequence(s) => {
-                let inner = resolve_type_name(types, s.type_param.id);
-                if inner == "u8" {
-                    "Bytes".to_string()
-                } else {
-                    format!("Vec<{}>", inner)
-                }
-            }
-            scale_info::TypeDef::Array(a) => {
-                format!("[{}; {}]", resolve_type_name(types, a.type_param.id), a.len)
-            }
-            scale_info::TypeDef::Tuple(t) => {
-                let inner: Vec<String> = t
-                    .fields
-                    .iter()
-                    .map(|f| resolve_type_name(types, f.id))
-                    .collect();
-                format!("({})", inner.join(", "))
-            }
-            _ => type_id.to_string(),
-        }
-    } else {
-        type_id.to_string()
-    }
 }
 
 /// Extract events from subxt's unified Metadata.
